增加全局调整按键透明度/边框和文字颜色
# Conflicts:
#	app/src/main/res/values-zh-rCN/strings.xml

+++ app/src/root/java/com.limelight/binding/input/evdev/EvdevCaptureProvider.kt
@@ -139,7 +139,7 @@ class EvdevCaptureProvider(
                                 // Other unhandled mouse buttons
                             }
                             else -> {
-                                val keyCode = EvdevTranslator.translateEvdevKeyCode(event.code)
+                                val keyCode = EvdevTranslator.translateEvdevKeyCode(event.code.toInt())
                                 if (keyCode.toInt() != 0) {
                                     listener.keyboardEvent(event.value != 0, keyCode)
                                 }
